add SNMOT 062

+++ label/label.py
@@ -1,6 +1,7 @@
+import logging
+
 import os
 import shutil
-import sys
 
 import cv2
 import numpy as np
@@ -13,6 +14,8 @@ from image import image
 PACKAGE_PATH = "label"
 SOCCER_NET_PATH = os.path.expandvars("${DATA_PATH}/soccer_net/tracking/train/")
 
+logging.basicConfig(format="%(levelname)s: %(message)s", level=logging.INFO)
+
 
 def key_to_label(key: str):
     if key == 112:  # key == "p"
@@ -23,6 +26,7 @@ def key_to_label(key: str):
 
 
 def load(filename: str, header: int = 0):
+    logging.info("loading %s", filename)
     with open(filename, "r", encoding="utf-8") as rfile:
         df = pd.read_csv(rfile, header=header)
 
@@ -30,23 +34,27 @@ def load(filename: str, header: int = 0):
 
 
 def save(filename: str, df: pd.DataFrame, header=True):
+    logging.info("saving %s", filename)
     with open(filename, "w", encoding="utf-8") as wfile:
         df.to_csv(wfile, index=False, header=header)
 
 
 def update_labels(filename: str):
+    logging.info("updating labels in %s", filename)
     data_dir = filename.split("_")[1]
     det_dir = os.path.join(SOCCER_NET_PATH, data_dir, "det")
     det_path = os.path.join(det_dir, "det.txt")
-    # Create a backup file
-    shutil.copy(det_path, os.path.join(det_dir, "det_back.txt"))
+
+    if not os.path.exists(os.path.join(det_dir, "det_back.txt")):
+        # Create a backup file
+        shutil.copy(det_path, os.path.join(det_dir, "det_back.txt"))
 
     # load both files
     det = load(det_path, header=None)
     df = load(os.path.join(PACKAGE_PATH, f"{filename}.csv"))
 
     # update labels
-    det["label"] = df["label"]
+    det["label"] = list(map(int, df["label"]))
 
     save(det_path, det, header=False)
 
@@ -55,7 +63,12 @@ def run():
 
     columns = ["image", "bbox", "label"]
     sn = SoccerNet(SOCCER_NET_PATH)
-    sn.collect()
+    sn.collect(["SNMOT-062"])
+
+    if len(sn.image_list) == 0:
+        raise ValueError(
+            "empty image list, did you set all environmental variables (DATA_DIR, REPO)?"
+        )
 
     file_dir = ""
     df = None
@@ -66,6 +79,8 @@ def run():
         new_file_dir = img_path.split("/")[0]
         if new_file_dir != file_dir:
             if filename:
+                # save labels to csv and update det.txt file
+                # in DATA_DIR when all images in a directory all labeled
                 save(filename, df)
                 update_labels(f"labels_{file_dir}")
 
@@ -95,14 +110,20 @@ def run():
                 key = cv2.waitKey(0)
                 if key == 27:  # esc
                     save(filename, df)
-                    sys.exit(0)
+                    return 0
 
                 key = key_to_label(key)
                 if key > 0:
                     break
-                print("invalid key, choose from {b - ball, p - player, esc - exit}")
+                logging.warning(
+                    "invalid key, choose from {b - ball, p - player, esc - exit}"
+                )
 
             img_data.append([img_path, i, key])
             cv2.destroyAllWindows()
 
         df = pd.concat([df, pd.DataFrame(img_data, columns=columns)])
+
+    # save update labels at the end
+    save(filename, df)
+    update_labels(f"labels_{file_dir}")
